requirements-dev.txt
# Include base requirements
-r requirements.txt

# Development and testing dependencies
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
django-stubs>=4.0.0

# Development tools
ipython>=8.0.0
django-extensions>=3.2.0

# Build and distribution
build>=0.10.0
twine>=4.0.0
